+++ CHANGES.rst
@@ -7,6 +7,7 @@ Unreleased
 - Change ugettext calls in models.py to ugettext_lazy
 - Fix history detail view on django-admin for abstract models (gh-308)
 - Fix issue with tracking an inherited model (abstract class) (gh-269)
+- Resolve issue where model references itself (gh-278)
 
 1.9.0 (2017-06-11)
 ------------------