+++ CHANGES.rst
@@ -8,6 +8,7 @@ Unreleased
 ----------
 - Use get_queryset rather than model.objects in history_view. (gh-303)
 - Change ugettext calls in models.py to ugettext_lazy
+- Resolve issue where model references itself (gh-278)
 
 1.9.0 (2017-06-11)
 ------------------